+++ src/Classes/MethodHelper.php
@@ -2,6 +2,7 @@
 
 namespace SebastiaanLuca\Helpers\Classes;
 
+use Error;
 use ReflectionException;
 use ReflectionMethod;
 
@@ -16,20 +17,18 @@ class MethodHelper
      *
      * @return bool
      */
-    public static function hasMethod($object, $method, $type)
+    public static function hasMethodOfType($object, $method, $type)
     {
-        // Primary check
         if (! method_exists($object, $method)) {
             return false;
         }
 
-        // Accessibility check
         try {
             $reflection = new ReflectionMethod($object, $method);
             $type = 'is' . studly_case($type);
 
             return $reflection->{$type}();
-        } catch (ReflectionException $exception) {
+        } catch (ReflectionException | Error $exception) {
             return false;
         }
     }
@@ -44,7 +43,7 @@ class MethodHelper
      */
     public static function hasProtectedMethod($object, $method)
     {
-        return self::hasMethod($object, $method, 'protected');
+        return static::hasMethodOfType($object, $method, 'protected');
     }
 
     /**
@@ -57,6 +56,6 @@ class MethodHelper
      */
     public static function hasPublicMethod($object, $method)
     {
-        return self::hasMethod($object, $method, 'public');
+        return static::hasMethodOfType($object, $method, 'public');
     }
 }